Handle a case when imgaug is not installed
