+++ albumentations/__init__.py
@@ -8,8 +8,8 @@ from .core.serialization import *
 from .augmentations import *
 
 try:
-    from .imgaug.transforms import *
+    from .imgaug.transforms import *  # type: ignore
 except ImportError:
-    # ImgAug is not installed by default, so we ignore this error.
+    # ImgAug is not installed by default, so we import stubs.
     # Run `pip install -U albumentations[imgaug] if you need augmentations from imgaug.`
-    pass
+    from .imgaug.stubs import *  # type: ignore
