+++ albumentations/imgaug/transforms.py
@@ -2,8 +2,8 @@ try:
     import imgaug as ia
 except ImportError as e:
     raise ImportError(
-        "You are trying to import a transform that uses the imgaug library, but imgaug is not installed. To install "
-        "a version of Albumentations that contains imgaug please run `pip install -U albumentations[imgaug]`"
+        "You are trying to import an augmentation that depends on the imgaug library, but imgaug is not installed. To "
+        "install a version of Albumentations that contains imgaug please run 'pip install -U albumentations[imgaug]'"
     ) from e
 
 try:
