+++ tests/files/transform_serialization_v2_with_totensor.json
