+++ tests/files/transform_serialization_v2_without_totensor.json
@@ -1,0 +1,1 @@
+{"__version__": "0.5.2", "transform": {"__class_fullname__": "Compose", "p": 1.0, "transforms": [{"__class_fullname__": "Blur", "always_apply": false, "p": 0.5, "blur_limit": [3, 7]}, {"__class_fullname__": "CLAHE", "always_apply": false, "p": 0.5, "clip_limit": [1, 4.0], "tile_grid_size": [8, 8]}, {"__class_fullname__": "ChannelDropout", "always_apply": false, "p": 0.5, "channel_drop_range": [1, 1], "fill_value": 0}, {"__class_fullname__": "ChannelShuffle", "always_apply": false, "p": 0.5}, {"__class_fullname__": "CoarseDropout", "always_apply": false, "p": 0.5, "max_holes": 8, "max_height": 8, "max_width": 8, "min_holes": 8, "min_height": 8, "min_width": 8, "fill_value": 0, "mask_fill_value": null}, {"__class_fullname__": "Downscale", "always_apply": false, "p": 0.5, "scale_min": 0.25, "scale_max": 0.25, "interpolation": 0}, {"__class_fullname__": "ElasticTransform", "always_apply": false, "p": 0.5, "alpha": 1, "sigma": 50, "alpha_affine": 50, "interpolation": 1, "border_mode": 4, "value": null, "mask_value": null, "approximate": false}, {"__class_fullname__": "Equalize", "always_apply": false, "p": 0.5, "mode": "cv", "by_channels": true}, {"__class_fullname__": "FancyPCA", "always_apply": false, "p": 0.5, "alpha": 0.1}, {"__class_fullname__": "Flip", "always_apply": false, "p": 0.5}, {"__class_fullname__": "GaussNoise", "always_apply": false, "p": 0.5, "var_limit": [10.0, 50.0], "per_channel": true, "mean": 0}, {"__class_fullname__": "GaussianBlur", "always_apply": false, "p": 0.5, "blur_limit": [3, 7], "sigma_limit": [0, 0]}, {"__class_fullname__": "GlassBlur", "always_apply": false, "p": 0.5, "sigma": 0.7, "max_delta": 4, "iterations": 2}, {"__class_fullname__": "GridDistortion", "always_apply": false, "p": 0.5, "num_steps": 5, "distort_limit": [-0.3, 0.3], "interpolation": 1, "border_mode": 4, "value": null, "mask_value": null}, {"__class_fullname__": "HorizontalFlip", "always_apply": false, "p": 0.5}, {"__class_fullname__": "HueSaturationValue", "always_apply": false, "p": 0.5, "hue_shift_limit": [-20, 20], "sat_shift_limit": [-30, 30], "val_shift_limit": [-20, 20]}, {"__class_fullname__": "ISONoise", "always_apply": false, "p": 0.5, "intensity": [0.1, 0.5], "color_shift": [0.01, 0.05]}, {"__class_fullname__": "ImageCompression", "always_apply": false, "p": 0.5, "quality_lower": 99, "quality_upper": 100, "compression_type": 0}, {"__class_fullname__": "InvertImg", "always_apply": false, "p": 0.5}, {"__class_fullname__": "LongestMaxSize", "always_apply": false, "p": 1, "max_size": 1024, "interpolation": 1}, {"__class_fullname__": "MedianBlur", "always_apply": false, "p": 0.5, "blur_limit": [3, 7]}, {"__class_fullname__": "MotionBlur", "always_apply": false, "p": 0.5, "blur_limit": [3, 7]}, {"__class_fullname__": "MultiplicativeNoise", "always_apply": false, "p": 0.5, "multiplier": [0.9, 1.1], "per_channel": false, "elementwise": false}, {"__class_fullname__": "NoOp", "always_apply": false, "p": 0.5}, {"__class_fullname__": "OpticalDistortion", "always_apply": false, "p": 0.5, "distort_limit": [-0.05, 0.05], "shift_limit": [-0.05, 0.05], "interpolation": 1, "border_mode": 4, "value": null, "mask_value": null}, {"__class_fullname__": "PadIfNeeded", "always_apply": false, "p": 1.0, "min_height": 1024, "min_width": 1024, "pad_height_divisor": null, "pad_width_divisor": null, "border_mode": 4, "value": null, "mask_value": null}, {"__class_fullname__": "Posterize", "always_apply": false, "p": 0.5, "num_bits": [4, 4]}, {"__class_fullname__": "RGBShift", "always_apply": false, "p": 0.5, "r_shift_limit": [-20, 20], "g_shift_limit": [-20, 20], "b_shift_limit": [-20, 20]}, {"__class_fullname__": "RandomBrightnessContrast", "always_apply": false, "p": 0.5, "brightness_limit": [-0.2, 0.2], "contrast_limit": [-0.2, 0.2], "brightness_by_max": true}, {"__class_fullname__": "RandomFog", "always_apply": false, "p": 0.5, "fog_coef_lower": 0.3, "fog_coef_upper": 1, "alpha_coef": 0.08}, {"__class_fullname__": "RandomGamma", "always_apply": false, "p": 0.5, "gamma_limit": [80, 120], "eps": null}, {"__class_fullname__": "RandomGridShuffle", "always_apply": false, "p": 0.5, "grid": [3, 3]}, {"__class_fullname__": "RandomRain", "always_apply": false, "p": 0.5, "slant_lower": -10, "slant_upper": 10, "drop_length": 20, "drop_width": 1, "drop_color": [200, 200, 200], "blur_value": 7, "brightness_coefficient": 0.7, "rain_type": null}, {"__class_fullname__": "RandomRotate90", "always_apply": false, "p": 0.5}, {"__class_fullname__": "RandomScale", "always_apply": false, "p": 0.5, "interpolation": 1, "scale_limit": [-0.09999999999999998, 0.10000000000000009]}, {"__class_fullname__": "RandomShadow", "always_apply": false, "p": 0.5, "shadow_roi": [0, 0.5, 1, 1], "num_shadows_lower": 1, "num_shadows_upper": 2, "shadow_dimension": 5}, {"__class_fullname__": "RandomSnow", "always_apply": false, "p": 0.5, "snow_point_lower": 0.1, "snow_point_upper": 0.3, "brightness_coeff": 2.5}, {"__class_fullname__": "RandomSunFlare", "always_apply": false, "p": 0.5, "flare_roi": [0, 0, 1, 0.5], "angle_lower": 0, "angle_upper": 1, "num_flare_circles_lower": 6, "num_flare_circles_upper": 10, "src_radius": 400, "src_color": [255, 255, 255]}, {"__class_fullname__": "Rotate", "always_apply": false, "p": 0.5, "limit": [-90, 90], "interpolation": 1, "border_mode": 4, "value": null, "mask_value": null}, {"__class_fullname__": "ShiftScaleRotate", "always_apply": false, "p": 0.5, "shift_limit_x": [-0.0625, 0.0625], "shift_limit_y": [-0.0625, 0.0625], "scale_limit": [-0.09999999999999998, 0.10000000000000009], "rotate_limit": [-45, 45], "interpolation": 1, "border_mode": 4, "value": null, "mask_value": null}, {"__class_fullname__": "SmallestMaxSize", "always_apply": false, "p": 1, "max_size": 1024, "interpolation": 1}, {"__class_fullname__": "Solarize", "always_apply": false, "p": 0.5, "threshold": [128, 128]}, {"__class_fullname__": "ToGray", "always_apply": false, "p": 0.5}, {"__class_fullname__": "ToSepia", "always_apply": false, "p": 0.5}, {"__class_fullname__": "Transpose", "always_apply": false, "p": 0.5}, {"__class_fullname__": "VerticalFlip", "always_apply": false, "p": 0.5}, {"__class_fullname__": "RandomCrop", "always_apply": false, "p": 1.0, "height": 32, "width": 32}, {"__class_fullname__": "Normalize", "always_apply": false, "p": 1.0, "mean": [0.485, 0.456, 0.406], "std": [0.229, 0.224, 0.225], "max_pixel_value": 255.0}], "bbox_params": null, "keypoint_params": null, "additional_targets": {}}}
